fix(brokers): change package names for brokers
move re brokers package to consumers
For RE broker, provide a publisher interface
Move alarms broker from consumer package
