+++ cmd/timescale-writer/main.go
@@ -15,12 +15,12 @@ import (
 	chclient "github.com/absmach/callhome/pkg/client"
 	consumertracing "github.com/absmach/magistrala/consumers/tracing"
 	httpapi "github.com/absmach/magistrala/consumers/writers/api"
+	"github.com/absmach/magistrala/consumers/writers/brokers"
 	"github.com/absmach/magistrala/consumers/writers/timescale"
 	"github.com/absmach/supermq"
 	"github.com/absmach/supermq/consumers"
 	smqlog "github.com/absmach/supermq/logger"
 	jaegerclient "github.com/absmach/supermq/pkg/jaeger"
-	"github.com/absmach/supermq/pkg/messaging/brokers"
 	brokerstracing "github.com/absmach/supermq/pkg/messaging/brokers/tracing"
 	pgclient "github.com/absmach/supermq/pkg/postgres"
 	"github.com/absmach/supermq/pkg/prometheus"
